refactor

+++ pipeline.py
@@ -2,26 +2,47 @@ import os
 import shutil
 import subprocess
 
-PENDING_PATH = "/mnt/c/Users/Mahkumazahn/Videos/Pending"
 TEMP_DIR = "./clips"
 
-# TODO: Make paths less nightmarish
+def main():
+    cleanup_last_run()
 
-print("moving files from pending")
-os.mkdir(TEMP_DIR)
+    print("moving files from pending")
+    os.mkdir(TEMP_DIR)
+    file_paths = copy_pending_files()
 
-filenames = os.listdir(PENDING_PATH)
-filepaths = [os.path.join(PENDING_PATH, filename) for filename in filenames]
-filepaths.sort(key=lambda filepath: os.path.getctime(filepath))
+    print("removing silence")
+    remove_silence(file_paths)
 
-for index, filepath in enumerate(filepaths):
-    # TODO: use shutil.move when pipeline is complete
-    shutil.copy(filepath, os.path.join(TEMP_DIR, str(index) + ".mp4"))
+def cleanup_last_run():
+    if os.path.exists(TEMP_DIR):
+        shutil.rmtree(TEMP_DIR)
 
-print("removing silence")
-clip_names = os.listdir(TEMP_DIR)
-for clip_name in clip_names:
-    os.system("./video-remove-silence/video-remove-silence " + os.path.join(TEMP_DIR, clip_name))
+def copy_pending_files():
+    # consider passing as argument
+    PENDING_PATH = "/mnt/c/Users/Mahkumazahn/Videos/Pending"
 
-shutil.rmtree(TEMP_DIR)
+    original_names = os.listdir(PENDING_PATH)
+    original_paths = [os.path.join(PENDING_PATH, original_name) for original_name in original_names]
+    original_paths.sort(key=lambda original_path: os.path.getctime(original_path))
 
+    print(original_paths)
+
+    new_paths = []
+    for index, original_path in enumerate(original_paths):
+        new_path = os.path.join(TEMP_DIR, str(index) + ".mp4")
+
+        # TODO: use shutil.move when pipeline is complete
+        shutil.copy(original_path, new_path)
+        new_paths.append(new_path)
+
+    return new_paths
+
+def remove_silence(file_paths):
+    for path in file_paths:
+        os.system("./video-remove-silence/video-remove-silence " + path)
+        os.remove(path)
+        os.rename(path.replace(".mp4", "_result.mp4"), path)
+        break
+
+main()
